acess collectMetricInterval atomically
change collectMetricInterval

1. access collectMetricInterval atomically
2. use the setting value for the KV MetricFlushInterval

+++ v1/ao/internal/reporter/reporter_test.go
@@ -265,7 +265,7 @@ func TestGRPCReporter(t *testing.T) {
 
 	assert.Equal(t, serviceKey, r.serviceKey)
 
-	assert.Equal(t, grpcMetricIntervalDefault, r.collectMetricInterval)
+	assert.Equal(t, int32(grpcMetricIntervalDefault), r.collectMetricInterval)
 	assert.Equal(t, grpcGetSettingsIntervalDefault, r.getSettingsInterval)
 	assert.Equal(t, grpcSettingsTimeoutCheckIntervalDefault, r.settingsTimeoutCheckInterval)
